pyproject.toml
[build-system]
requires = ["poetry-core>=1.9.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[project]
name = "jira-extractor-api-first"
version = "0.1.0"
description = "Jira extractor with API-first strategy and Playwright fallback"
readme = "README.md"
requires-python = ">=3.11,<3.14"
authors = [{ name = "Data Engineering" }]
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn[standard]>=0.30.0,<1.0.0",
  "requests>=2.32.0,<3.0.0",
  "pandas>=2.2.0,<3.0.0",
  "pyarrow>=17.0.0,<20.0.0",
  "python-dotenv>=1.0.0,<2.0.0",
  "pydantic>=2.8.0,<3.0.0",
  "pydantic-settings>=2.3.0,<3.0.0",
  "tenacity>=9.0.0,<10.0.0",
  "playwright>=1.50.0,<2.0.0",
  "black (>=26.1.0,<27.0.0)",
]

[project.scripts]
extractor-run = "extractor.run:main"
extractor-api = "api.main:run"

[tool.poetry]
packages = [{ include = "extractor" }, { include = "api" }]

[tool.poetry.group.dev.dependencies]
pytest = ">=8.3.0,<9.0.0"
pytest-cov = ">=5.0.0,<6.0.0"
httpx = ">=0.27.0,<1.0.0"

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
